requirements.txt
streamlit>=1.30.0
transformers>=5.0.0
torch>=2.0.0
pillow>=10.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
pdf2image>=1.16.0
